- Added Backup Seed Activity to show the mnemonic for the recently created seed

+++ app/src/main/java/cy/agorise/crystalwallet/viewmodels/AccountSeedViewModel.java
@@ -29,7 +29,7 @@ public class AccountSeedViewModel extends AndroidViewModel {
         this.app = application;
     }
 
-    public void loadSeed(int seedId){
+    public void loadSeed(long seedId){
         this.accountSeed = this.db.accountSeedDao().findByIdLiveData(seedId);
     }
